Reorganized file structure for app and rebuilt paths. Added blog images to home page below carousel.
