--- src/Authors.css
@@ -1,27 +1,0 @@
-/* Styles for Authors page*/
-.authors {
-  border-radius: 50%;
-  margin-left: 30px;
-  margin-bottom: 5%;
-  width: 200px;
-  height: 300px;
-  content: wrap;
-}
-.transparent {
-  border: none;
-}
-.profile-bg {
-  background-color: rgb(119, 119, 119);
-  color: white;
-}
-
-.AuthorHeader,
-.Accordion {
-  margin-top: 4%;
-  margin-bottom: 2%;
-}
-
-.Accordion {
-  margin-bottom: 10%;
-}
-/* Styles for Authors Page*/
