+++ src/Components/Navbar.js
@@ -1,6 +1,5 @@
 import React from "react";
 import { useLocation } from "react-router-dom";
-import "./index.css";
 import { Navbar, Nav, NavDropdown } from "react-bootstrap";
 
 const Navigation = (props) => {
