--- src/HomePage.js
@@ -1,12 +1,0 @@
-import React from "react";
-import { Slideshow } from "./Carousel";
-
-const HomePage = () => {
-  return (
-    <>
-      <Slideshow />
-    </>
-  );
-};
-
-export default HomePage;
